+++ test/ninja/mbedded/ninjaterm/integration/NinjaTermAppTest.java
@@ -5,10 +5,10 @@ import javafx.application.Platform;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
-import javafx.scene.text.TextFlow;
 import javafx.stage.Stage;
-import ninja.mbedded.ninjaterm.managers.ComPortManager;
 import ninja.mbedded.ninjaterm.model.Model;
+import ninja.mbedded.ninjaterm.util.comport.ComPort;
+import ninja.mbedded.ninjaterm.util.comport.ComPortFactory;
 import ninja.mbedded.ninjaterm.view.mainWindow.MainWindowViewController;
 import org.controlsfx.glyphfont.GlyphFont;
 import org.junit.Test;
@@ -16,13 +16,11 @@ import org.testfx.framework.junit.ApplicationTest;
 
 import java.io.IOException;
 
-import static javafx.scene.input.KeyCode.ENTER;
 import static junit.framework.TestCase.assertEquals;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.*;
 import static org.testfx.api.FxAssert.verifyThat;
-import static org.testfx.matcher.base.NodeMatchers.hasChildren;
 import static org.testfx.matcher.base.NodeMatchers.hasText;
-
-import org.testfx.service.finder.*;
 
 /**
  * Created by gbmhu on 2016-10-28.
@@ -36,10 +34,21 @@ public class NinjaTermAppTest extends ApplicationTest {
     @Override
     public void start(Stage stage) {
 
+        //==============================================//
+        //==== MOCK COMPORT FACTORY AND COM PORT =======//
+        //==============================================//
+
         GlyphFont glyphFont = new GlyphFont("FontAwesome", 12, "/ninja/mbedded/ninjaterm/resources/fontawesome-webfont.ttf");
 
+        ComPort comPort = mock(ComPort.class);
+        when(comPort.scan()).thenReturn(new String[]{"COM1", "COM2"});
+
+        ComPortFactory comPortFactory = mock(ComPortFactory.class);
+        when(comPortFactory.create()).thenReturn(comPort);
+
+
         // Create application model (data/state)
-        model = new Model();
+        model = new Model(comPortFactory);
 
         FXMLLoader loader = new FXMLLoader(getClass().getResource("/ninja/mbedded/ninjaterm/view/mainWindow/MainWindowView.fxml"));
         try {
@@ -49,7 +58,7 @@ public class NinjaTermAppTest extends ApplicationTest {
         }
         mainWindowViewController = loader.getController();
 
-        mainWindowViewController.init(model, glyphFont, new ComPortManager());
+        mainWindowViewController.init(model, glyphFont);
 
         //mainWindowViewController.addNewTerminal();
 
@@ -69,8 +78,9 @@ public class NinjaTermAppTest extends ApplicationTest {
             //rightClickOn("#mainVBox").moveTo("New").clickOn("Text Document");
             //write("myTextfile.txt").push(ENTER);
 
-            // Make sure the open/close COM port button says "Open" by default
+            // Make sure the open/close COM port button says "Open" by default, and is currently disabled
             verifyThat(mainWindowViewController.statusBarViewController.openCloseComPortButton, hasText("Open"));
+            //assertEquals(true, mainWindowViewController.statusBarViewController.openCloseComPortButton.isDisable());
 
             // Make sure there is one terminal by default
             assertEquals(1, mainWindowViewController.terminalViewControllers.size());
@@ -78,9 +88,12 @@ public class NinjaTermAppTest extends ApplicationTest {
             //clickOn();
 
             int numOfChildrenBeforeAction = mainWindowViewController.statusBarViewController.statusTextFlow.getChildren().size();
+
+            // Hit the scan button
             mainWindowViewController.terminalViewControllers.get(0).comSettingsViewController.reScanButton.fire();
 
             assertEquals(numOfChildrenBeforeAction + 1, mainWindowViewController.statusBarViewController.statusTextFlow.getChildren().size());
+
 
             // Click on the button
             //clickOn(mainWindowViewController.terminalViewControllers);